feat(dashboard): add TableListSettingsShell component for enhanced settings management in dashlets
